requirements.txt
openai
python-dotenv
Pillow
reportlab
